what: fix zip creation logic, update help text
Fixed: Running both '-m' and '-z' would result in the creation of two identical zips (since modules don't get automatically removed).
Fixed: Running '-a' by itself would do nothing. The new behavior demands that, if '-m' or '-a' are to be used by themselves without '-z', nothing happens.
Fixed: The parameter '-a' was useless because the packing script defaulted to signing the script regardless of input. Prepare for a fix of the script in the kerneller repo.